Strangely, groff accepts .ll arguments with multiple signs.  For odd numbers of minus signs, the intended behaviour is decreasing the line length, for even numbers, increasing it.  The code in term_setwidth() resulted in incorrect behaviour in two cases: for more than two signs, the line length wasn't changed at all because a2roffsu() was called incorrectly and failed, and if the second sign was negative, a negative width was passed to the setwidth() callbacks, which they aren't prepared to handle.
Fix this by iterating over all signs to find the correct iop operation
code (0=absolute, 1=increase, -1=decrease).  Also improve code clarity
by making the width argument of the setwidth() callbacks unsigned and
removing some ugly casts in these callbacks.

+++ term.h
@@ -111,7 +111,7 @@ struct	termp {
 	void		(*end)(struct termp *);
 	void		(*endline)(struct termp *);
 	void		(*advance)(struct termp *, size_t);
-	void		(*setwidth)(struct termp *, int, int);
+	void		(*setwidth)(struct termp *, int, size_t);
 	size_t		(*getwidth)(const struct termp *, int);
 	int		(*hspan)(const struct termp *,
 				const struct roffsu *);
