+++ term_ascii.c
@@ -49,7 +49,7 @@ static	void		  ascii_begin(struct termp *);
 static	void		  ascii_end(struct termp *);
 static	void		  ascii_endline(struct termp *);
 static	void		  ascii_letter(struct termp *, int);
-static	void		  ascii_setwidth(struct termp *, int, int);
+static	void		  ascii_setwidth(struct termp *, int, size_t);
 
 #if HAVE_WCHAR
 static	void		  locale_advance(struct termp *, size_t);
@@ -152,14 +152,14 @@ locale_alloc(const struct manoutput *outopts)
 }
 
 static void
-ascii_setwidth(struct termp *p, int iop, int width)
+ascii_setwidth(struct termp *p, int iop, size_t width)
 {
 	p->tcol->rmargin = p->defrmargin;
 	if (iop > 0)
 		p->defrmargin += width;
 	else if (iop == 0)
-		p->defrmargin = width ? (size_t)width : p->lastrmargin;
-	else if (p->defrmargin > (size_t)width)
+		p->defrmargin = width ? width : p->lastrmargin;
+	else if (p->defrmargin > width)
 		p->defrmargin -= width;
 	else
 		p->defrmargin = 0;
