+++ term_ps.c
@@ -110,7 +110,7 @@ static	void		  ps_printf(struct termp *, const char *, ...)
 				__attribute__((__format__ (__printf__, 2, 3)));
 static	void		  ps_putchar(struct termp *, char);
 static	void		  ps_setfont(struct termp *, enum termfont);
-static	void		  ps_setwidth(struct termp *, int, int);
+static	void		  ps_setwidth(struct termp *, int, size_t);
 static	struct termp	 *pspdf_alloc(const struct manoutput *, enum termtype);
 static	void		  pdf_obj(struct termp *, size_t);
 
@@ -623,7 +623,7 @@ pspdf_alloc(const struct manoutput *outopts, enum termtype type)
 }
 
 static void
-ps_setwidth(struct termp *p, int iop, int width)
+ps_setwidth(struct termp *p, int iop, size_t width)
 {
 	size_t	 lastwidth;
 
@@ -631,8 +631,8 @@ ps_setwidth(struct termp *p, int iop, int width)
 	if (iop > 0)
 		p->ps->width += width;
 	else if (iop == 0)
-		p->ps->width = width ? (size_t)width : p->ps->lastwidth;
-	else if (p->ps->width > (size_t)width)
+		p->ps->width = width ? width : p->ps->lastwidth;
+	else if (p->ps->width > width)
 		p->ps->width -= width;
 	else
 		p->ps->width = 0;
